pyproject.toml
[project]
name = "pyscim"
version = "0.1.0"
description = "Standards-compliant SCIM 2.0 server implementation"
readme = "README.md"
authors = [
    { name = "Nathan Catania", email = "nathan@nathancatania.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "tortoise-orm[asyncpg]>=0.21.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.0",
    "supabase>=2.10.0",
    "httpx>=0.27.0",
    "email-validator>=2.2.0",
    "pyparsing>=3.2.0",
    "click>=8.1.0",
]

[project.scripts]
pyscim = "scim.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scim"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
